Split TimelockAuthorizer into multiple files
